+++ ProjMain/Assets/Prefabs/MovingObejcts.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 22c416aef29694743b0dcb1b074105a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
