--- ProjMain/Assets/Prefabs/SpawnObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 281b2177eae700543aa1e5e4c2699939
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
