+++ ProjMain/Assets/Prefabs/Trap/14_0.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 96f42dff3a8a23e48808998e9418270e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
