+++ ProjMain/Assets/Prefabs/Trap/BearTrap.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5194a730fbf08a74687217adb5bf47d3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
