+++ ProjMain/Assets/Prefabs/Trap/CircleTrap.prefab
@@ -14,8 +14,8 @@ GameObject:
   - component: {fileID: 1851061488032066441}
   m_Layer: 0
   m_Name: CircleTrap
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
+  m_TagString: Trap
+  m_Icon: {fileID: 5132851093641282708, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
