+++ ProjMain/Assets/Prefabs/Trap/FireTrap.prefab
@@ -14,8 +14,8 @@ GameObject:
   - component: {fileID: 506631639320425706}
   m_Layer: 0
   m_Name: FireTrap
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
+  m_TagString: Trap
+  m_Icon: {fileID: 3936346786652291628, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
@@ -136,7 +136,7 @@ BoxCollider2D:
   m_CallbackLayers:
     serializedVersion: 2
     m_Bits: 4294967295
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0.40949738, y: 0}
