+++ ProjMain/Assets/Prefabs/Trap/SpawnObject.cs
@@ -1,0 +1,39 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SpawnObject : MonoBehaviour
+{
+    // Start is called before the first frame update
+    public GameObject Spawn;
+    CapsuleCollider2D capsule;
+   
+    void Start()
+    {
+        capsule = GetComponent<CapsuleCollider2D>();
+       
+        Spawner();
+    }
+    void Spawner()
+    {
+        int Rand = Random.Range(0, 10);
+        if (Rand < 6)
+        {
+            Instantiate(Spawn, transform.position, transform.rotation);
+            capsule.enabled = true;
+            
+        }
+        else { 
+        capsule.enabled = false;
+      
+    }
+
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
