+++ ProjMain/Assets/Prefabs/Trap/SpawnObjectBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aee9d0a7addfd144dbff6a462dc22076
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
