--- ProjMain/Assets/Prefabs/Trap/Trap.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5194a730fbf08a74687217adb5bf47d3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
