+++ ProjMain/Assets/Prefabs/Trap/small_wood_spike_04 (11).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3e094d42b2bdb2e4784e90c148e32509
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
