--- ProjMain/Assets/craftpix/Traps/TrapGuns/14_0.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 96f42dff3a8a23e48808998e9418270e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
