+++ ProjMain/Logs/shadercompiler-AssetImportWorker0.log
@@ -1,4 +1,4 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2022.3.6f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2022.3.6f1/Editor/Data/PlaybackEngines'
+Base path: 'D:/Unity/2022.3.6f1/Editor/Data', plugins path 'D:/Unity/2022.3.6f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
 Unhandled exception: Protocol error - failed to read magic number (error -2147483644, transferred 0/4)
